Move the search logic from ProductList to Header
