+++ e-keithel-frontend/src/app/components/ProductList.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { fetchProductList } from '../../redux/reducers/product/productReducers';
 import Product from './Product';
@@ -10,22 +10,9 @@ const ProductList = () => {
     (state) => state.persistedReducer.productListReducer.productList.data
   );
 
-  const search = useSelector((state) => state.persistedReducer.searchReducer);
-  console.log('search ', search.searchWord);
-
   useEffect(() => {
     dispatch(fetchProductList());
   }, [dispatch]);
-
-  const [searchResults, setSearchResults] = useState([]);
-
-  useEffect(() => {
-    const searchResults = productList.filter((product) =>
-      product.title.toLowerCase().includes(search.searchWord.toLowerCase())
-    );
-    console.log('search results: ', searchResults);
-    setSearchResults(searchResults);
-  }, [search]);
 
   return (
     <div className="productList">
@@ -35,7 +22,7 @@ const ProductList = () => {
         </div>
 
         <div className="productList__row">
-          {searchResults.map((product) => {
+          {productList.map((product) => {
             return (
               <Product
                 key={product.id}
